Allow multiple, optional `devise` `authentication_keys`
In the `devise` instructions for using multiple `authentication_keys`, in order to make them optional, you need to use a hash with `false` or `true` values (optional vs. required). The way that `devise_token_auth` compares `authentication_keys` assumes an array. This changes to using a hash.

NOTE: it breaks it using an array though :'(

https://github.com/plataformatec/devise/wiki/How-To:-Allow-users-to-sign-in-using-their-username-or-email-address#tell-devise-to-use-login-in-the-authentication_keys